services/auth/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deep-analysis-auth"
version = "0.4.0.dev0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "structlog>=24.2",
    "prometheus-client>=0.20",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "asyncpg>=0.29",
    "argon2-cffi>=23.1",
    "pyjwt[crypto]>=2.8",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "cryptography>=42.0",
]

[tool.hatch.build.targets.wheel]
packages = ["auth_service"]
